redo module header, works with Rack
