+++ modules/inc/delay.hpp
@@ -66,8 +66,6 @@ public:
 
 };
 
-#ifdef BUILD_F373
-
 /// Macro used to specify the number of samples to per DAC transfer.
 #define DELAY_BUFFER_SIZE 1
 
@@ -76,7 +74,7 @@ void delayTouchLink (void *);
 
 /// Calibration/template module class.
 /** A simple self calibration tool that doubles as an introductory template.*/
-class ViaDelay : public ViaModule {
+class ViaDelay : public TARGET_VIA {
 
 public:
 
@@ -269,7 +267,6 @@ public:
 		render(DELAY_BUFFER_SIZE);
 	}
 	void slowConversionCallback(void) {
-		LOGICA_HIGH;
 		upsamplesLeft = 8;
 		controls.updateExtra();
 		int32_t rawControls = 4095 - __USAT(controls.cv1Value - 2048 + controls.knob1Value, 12);
@@ -277,7 +274,6 @@ public:
 		delayTimeTarget = revExpo.convert(rawControls);
 		delayTimeTarget = fix16_mul(delayTimeTarget, 160000);
 		slewFactor = (delayTimeTarget - lastTarget) >> 3;
-		LOGICA_LOW;
 
 	}
 	void auxTimer1InterruptCallback(void) {
@@ -310,7 +306,5 @@ public:
 	}
 
 };
-
-#endif
 
 #endif /* INC_Calib_HPP_ */
