+++ modules/inc/sync.hpp
@@ -20,9 +20,6 @@ class PllController {
 	int32_t lastMultiplier = 1;
 	int32_t lastYIndex = 0;
 
-#ifdef BUILD_F373
-	int32_t lastRatioX = 1;
-#endif
 	int32_t ratioXTransitionPoint = 0;
 	int32_t ratioXStable = 1;
 
@@ -41,9 +38,6 @@ class PllController {
 
 	}
 
-#ifdef BUILD_F373
-	int32_t lastRatioY = 1;
-#endif
 	int32_t ratioYTransitionPoint = 0;
 	int32_t ratioYStable = 1;
 
@@ -66,10 +60,10 @@ class PllController {
 
 public:
 
-#ifdef BUILD_VIRTUAL
 	int32_t lastRatioX = 1;
 	int32_t lastRatioY = 1;
-#endif
+
+
 
 	uint32_t virtualTimer;
 
@@ -233,7 +227,7 @@ enum sync_aux4Modes {groupSpecific, global};
 
 void syncTouchLink (void *);
 
-class ViaSync : public ViaModule {
+class ViaSync : public TARGET_VIA {
 
 public:
